+++ src/mocks/handler.ts
@@ -1,6 +1,5 @@
 import { http, HttpResponse, PathParams } from "msw";
 import { LOGIN_END_POINT, SIGN_UP_END_POINT } from "@/features/auth/constants";
-import { validateNickname } from "@/features/auth/lib";
 import { MarkingListRequest } from "@/features/marking/api";
 import { MARKING_REQUEST_URL } from "@/features/marking/constants";
 // data
@@ -291,6 +290,175 @@ export const profileHandlers = [
   ),
 ];
 
+export const addressHandlers = [
+  http.get("http://localhost/addresses", (req) => {
+    const {
+      request: { url },
+    } = req;
+
+    const URLObject = new URL(url);
+    const keyword = URLObject.searchParams.get("keyword");
+
+    if (keyword === "강남구 역삼동") {
+      return HttpResponse.json({
+        code: 200,
+        message: "good",
+        content: [
+          {
+            id: 0,
+            province: "서울특별시",
+            cityCounty: "강남구",
+            subDistrict: "역삼1동",
+            district: "123-45",
+          },
+          {
+            id: 1,
+            province: "서울특별시",
+            cityCounty: "강남구",
+            subDistrict: "역삼2동",
+            district: "123-45",
+          },
+          {
+            id: 2,
+            province: "서울특별시",
+            cityCounty: "강남구",
+            subDistrict: "역삼3동",
+            district: "123-45",
+          },
+          {
+            id: 3,
+            province: "서울특별시",
+            cityCounty: "강남구",
+            subDistrict: "역삼4동",
+            district: "123-45",
+          },
+          {
+            id: 4,
+            province: "서울특별시",
+            cityCounty: "강남구",
+            subDistrict: "역삼5동",
+            district: "123-45",
+          },
+          {
+            id: 5,
+            province: "서울특별시",
+            cityCounty: "강남구",
+            subDistrict: "역삼6동",
+            district: "123-45",
+          },
+          {
+            id: 6,
+            province: "서울특별시",
+            cityCounty: "강남구",
+            subDistrict: "역삼7동",
+            district: "123-45",
+          },
+        ],
+      });
+    }
+
+    if (keyword === "도봉구 도봉동") {
+      return HttpResponse.json({
+        code: 200,
+        message: "good",
+        content: [
+          {
+            id: 0,
+            province: "서울특별시",
+            cityCounty: "도봉구",
+            subDistrict: "도봉1동",
+            district: "123-45",
+          },
+          {
+            id: 1,
+            province: "서울특별시",
+            cityCounty: "도봉구",
+            subDistrict: "도봉2동",
+            district: "123-45",
+          },
+          {
+            id: 2,
+            province: "서울특별시",
+            cityCounty: "도봉구",
+            subDistrict: "도봉3동",
+            district: "123-45",
+          },
+          {
+            id: 3,
+            province: "서울특별시",
+            cityCounty: "도봉구",
+            subDistrict: "도봉4동",
+            district: "123-45",
+          },
+        ],
+      });
+    }
+
+    return HttpResponse.json({
+      code: 204, // 검색 결과 없을 시를 가정
+      message: "bad",
+      content: [],
+    });
+  }),
+  http.get("http://localhost/addresses/search-by-location", () => {
+    return HttpResponse.json({
+      code: 200,
+      message: "good",
+      content: [
+        {
+          id: 0,
+          province: "서울특별시",
+          cityCounty: "영등포구",
+          subDistrict: "영등포 1가",
+          district: "123-45",
+        },
+        {
+          id: 1,
+          province: "서울특별시",
+          cityCounty: "영등포구",
+          subDistrict: "영등포 2가",
+          district: "123-45",
+        },
+        {
+          id: 2,
+          province: "서울특별시",
+          cityCounty: "영등포구",
+          subDistrict: "영등포 3가",
+          district: "123-45",
+        },
+        {
+          id: 3,
+          province: "서울특별시",
+          cityCounty: "영등포구",
+          subDistrict: "영등포 4가",
+          district: "123-45",
+        },
+        {
+          id: 4,
+          province: "서울특별시",
+          cityCounty: "영등포구",
+          subDistrict: "영등포 5가",
+          district: "123-45",
+        },
+        {
+          id: 5,
+          province: "서울특별시",
+          cityCounty: "영등포구",
+          subDistrict: "영등포 6가",
+          district: "123-45",
+        },
+        {
+          id: 6,
+          province: "서울특별시",
+          cityCounty: "영등포구",
+          subDistrict: "영등포 7가",
+          district: "123-45",
+        },
+      ],
+    });
+  }),
+];
+
 // * 나중에 msw 사용을 대비하여 만들었습니다.
 export const handlers = [
   ...signUpByEmailHandlers,
@@ -298,4 +466,5 @@ export const handlers = [
   ...markingModalHandlers,
   ...loginHandlers,
   ...profileHandlers,
+  ...addressHandlers,
 ];
